Clarify where applications should be run in howtos
The wording is adjusted to make it clearer when one pear app should be
left running and where to run apps in general. In some cases a reader
would be prompted to make a directory and then change directories into
it and modify some files. After modifying the files, they would be asked
to change into the directory again even though they would be in the
directory already. This has been clarified by removing commands for
changing directory where the reader will be already in that directory.