Update Live to 1.11.x (#172)
* Clarify where to run apps in howtos (#170)

* Add missing argument for `drive-writer-app` key in p2p fs howto

* Clarify where applications should be run in howtos

The wording is adjusted to make it clearer when one pear app should be
left running and where to run apps in general. In some cases a reader
would be prompted to make a directory and then change directories into
it and modify some files. After modifying the files, they would be asked
to change into the directory again even though they would be in the
directory already. This has been clarified by removing commands for
changing directory where the reader will be already in that directory.

* remove-redundant-replicate-corestore (#171)

* Pear release 1.11.1 docs (#169)

* Add `data` cli command & info about `apps` subcommand

* Add `pear reset` cli command for resetting a Pear application

Added a warning in the description as the command can result in lost
data and users should be cautious when using it.

* Remove leading spaces for `data` command flags

* Add note about dumping paths from `pear://` links in `dump` command

* Add `Pear.tray()` to Pear API docs

* Add supported formats for tray icon

* Remove `--unsafe-clear-preferences` from `pear run` cli doc

* Add `pear.gui.name` config option

* Edit new documentation for grammar & formatting

* Clarify `pear.gui.name` is an override & doesnt have a default

* Remove using "you" in `pear dump` doc & Correct grammar in `Pear.tray()`

Co-authored-by: David Mark Clements <huperekchunow@googlemail.com>

* Reword `pear reset` to make warning explicit

* Add note about hidden app issues when linux doesn't support `Pear.tray()`

---------

Co-authored-by: David Mark Clements <huperekchunow@googlemail.com>

---------

Co-authored-by: Marco <maidh91@gmail.com>
Co-authored-by: David Mark Clements <huperekchunow@googlemail.com>